feat(resolver): simplify semantic convention registry resolution function
